pyproject.toml
[project]
name = "number-quiz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "kanjize>=1.6.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "wanakana-python>=1.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.12.5",
]
